Update openapi.yaml #34
The differences are:

The HTTP method has been changed from post to get.
The requestBody section has been replaced with a parameters section. The filePaths parameter is now a query parameter (in: query) instead of being part of the request body. This is because GET requests should not have a body.